feat: changed Feeding delegate
