+++ src/TeachMeSkills.Shchypakin.Homework-5/Program.cs
@@ -9,9 +9,9 @@ namespace TeachMeSkills.Shchypakin.Homework_5
     {
         static void Main(string[] args)
         {
-            Chicken chicken = new Chicken("Ch");
-            Chicken chicken1 = new Chicken("Ch1");
-            Chicken chicken2 = new Chicken("Ch2");
+            Chicken chicken = new Chicken("Chicken1");
+            Chicken chicken1 = new Chicken("Chicken2");
+            Chicken chicken2 = new Chicken("Chicken3");
             Seal seal1 = new Seal("Seal1");
             Elephant elephant1 = new Elephant("Elephant1");
             Tiger tiger1 = new Tiger("Tiger1");
